requirements.txt
blinker==1.9.0
captcha==0.7.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
factorial==1.0.0
Flask==3.1.3
flask-cors==6.0.2
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpath==0.82.2
loguru==0.7.3
lxml==6.0.2
MarkupSafe==3.0.3
pillow==11.3.0
PyJWT==2.11.0
requests==2.32.5
urllib3==2.5.0
vboxapi==1.0
Werkzeug==3.1.6
win32_setctime==1.2.0
zipp==3.23.0
gunicorn
